requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
Django==4.1.4
idna==3.4
liqpay-sdk-python3==1.0.3
Pillow==9.3.0
pycparser==2.21
pyOpenSSL==23.0.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
